requirements.txt
langchain-community
langchain-openai
langchain-pinecone
pinecone
python-dotenv